+++ wxviews/binding/__init__.py
@@ -1,4 +1,3 @@
 """Binding implementation"""
 
-from .event import EventBinding, TextTwoWaysRule, CheckBoxTwoWaysRule
 from .modifiers import bind